New: usages of Entities and formates into dishes.html

+++ App/src/Models/Entities/Dishes/Dish.php
@@ -5,9 +5,9 @@ namespace App\src\Models\Entities\Dishes;
 class Dish {
     private string $name;
     private string $description;
-    private int $prix;
+    private float $prix;
 
-    public function __construct(string $name, string $description, int $prix) {
+    public function __construct(string $name, string $description, float $prix) {
         $this->name = $name;
         $this->description = $description;
         $this->prix = $prix;
@@ -21,7 +21,7 @@ class Dish {
         return $this->description;
     }
 
-    public function getPrix(): int {
+    public function getPrix(): float {
         return $this->prix;
     }
 }
